Move Channel ABC from __init__.py to dedicated base module
